[bug fix] NeuralNet::sort_connections() was not being called  due to an uninitialized member (connections_sorted) which caused sort_connections() to prematurally return. After resolving this issue, it was discovvered that the used[] vector was not being utilized to prevent repetitive discovery of connections in the sorting. This was also fixed and sorting is now correct.
Note: the first evaluation of a recurrent connection of a network will return sigma(0) as an input to
the connections destination. Initially we thought that an input of 0 would make more sense, but in
retrospect, its not so clear which is better. For now we will stick with sigma(0), but it may be something
to revisit in the future.