Prior commit: Innovation for user added input/bias/output nodes has been changed to the hash of an incrementing/zero/decrementing integer and the previous node hash. The gene innovation number remains the same, but origin and destination are now the innovation hash of the node, not it's position in the list. With seperate NodeGene and ConnectionGene innovation, we can add together structural differences between two parents, instead of only taking structure from the more fit parent (as in NEAT). We can also default to doing standard NEAT crossover by simply setting Probabilities::single_greater =1.0 and single_lesser = 0.0.
This commit:
Instead of an unsorted_map for the connection_genes, I utilize a container that utilizes both a unsorted_map and a vector,
insertion_ordered_map. This was required because when building the neural network, the would_make_loop(i,j) function would
incorrectly label a forward connection as recurrent, if the recurrent connection was added to the network first. This
happened because looping over an unsorted_map does not guarantee any order. A std::map is one solution, but I chose instead
to retain the constant time look up speed and to carry an additional internal vector in the insertion_ordered_map so that
the order of insertion is easily tracked and the vector is cached.

In addition I have added a GeneticDistance function which calculates the NEAT delta of two genomes. In our case we do not
distinguish between excess and disjoint genes. This is likely unnecessary (since in NEAT c1=c2=1.0) and also not easily
possible in our implementation, since we maintain no global state (innovation number), but instead utilize hashing functions
to determine equivalence. Added tests using GeneticDistance and new crossover.