Refactored ConcurrentNeuralNet::build_action_list to follow the lifetime of each node:
start | origin of R | zero-out | dest of R / dest of N | sigmoid | origin of N | end

In addition, added check on compare_connections for the case of the self-recurrent
connection, which needs to happen before any zeroing out takes place, and is thus
promoted to be evaluated before all other connections.

For future load balancing, the earliest and latest possible set for each node action
is tracked.

ConsecutiveNeuralNet: added a line in sort_connections to ensure that all nodes are
put in the sigmoided state from the very beginning so that during the first evaluation
empty nodes are not sigmoided before being used as the origin of recurrent connections.