Refactored NeuralNet and subclasses. Now implement intermediate abstract class between the NeuralNet base class and the seperate implementations. The intermediate layer follows the Curiously Recursive Template Pattern.
Now implement a GenomeConverter type erasure struct for storing the target NeuralNet
type for libNeat. This is private to class Population. One side effect of this poly-
morphism is that Organisms now has to hold a std::unique_ptr<NeuralNet> instead of
holding it by value. This means the NeuralNets are spread out on the heap. This may
affect performance slightly. Needs investigation.

Python bindings are not fully updated to new layout and require attention.